refactor(calendar) : add utils for generate date and delete onclick on date
